it is asked by Bert to commit the changes for solving compile error. it needs someone to check this revision.
git-svn-id: https://svn.opendtect.org/od/trunk@38545 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/MPEEngine/horizonadjuster.cc
@@ -220,7 +220,7 @@ bool HorizonAdjuster::track( const BinID& from, const BinID& to,
 	    return false;
 
 	const bool res = tracker_->track();
-	const float resz = zsamp.atIndex( tracker_->targetDepth() );
+	const float resz = zsamp.atIndex( (int)tracker_->targetDepth() );
 	if ( !permittedZRange().includes(resz-startz,false) )
 	    return false;
 
@@ -233,7 +233,7 @@ bool HorizonAdjuster::track( const BinID& from, const BinID& to,
 	return false;
 
     const bool res = tracker_->track();
-    const float resz = zsamp.atIndex( tracker_->targetDepth() );
+    const float resz = zsamp.atIndex( (int)tracker_->targetDepth() );
     if ( !permittedZRange().includes(resz-startz,false) )
 	return false;
